Move benchmark_wrappers to benchmark folder.
PiperOrigin-RevId: 306521269
