+++ official/benchmark/bert_squad_benchmark.py
@@ -33,7 +33,7 @@ from official.benchmark import bert_benchmark_utils as benchmark_utils
 from official.nlp.bert import run_squad
 from official.utils.misc import distribution_utils
 from official.utils.misc import keras_utils
-from official.utils.testing import benchmark_wrappers
+from official.benchmark import benchmark_wrappers
 
 
 # pylint: disable=line-too-long
