+++ official/benchmark/keras_cifar_benchmark.py
@@ -23,7 +23,7 @@ from absl import flags
 import tensorflow as tf  # pylint: disable=g-bad-import-order
 
 from official.benchmark import keras_benchmark
-from official.utils.testing import benchmark_wrappers
+from official.benchmark import benchmark_wrappers
 from official.benchmark.models import resnet_cifar_main
 
 MIN_TOP_1_ACCURACY = 0.929
