+++ official/benchmark/resnet_ctl_imagenet_benchmark.py
@@ -25,7 +25,7 @@ import tensorflow as tf
 from official.vision.image_classification.resnet import common
 from official.vision.image_classification.resnet import resnet_ctl_imagenet_main
 from official.benchmark.perfzero_benchmark import PerfZeroBenchmark
-from official.utils.testing import benchmark_wrappers
+from official.benchmark import benchmark_wrappers
 from official.utils.flags import core as flags_core
 
 MIN_TOP_1_ACCURACY = 0.76
